requirements.txt
graphene==2.1.3
flask-graphql==2.0.0
python-whois==0.7.1
redis==3.0.1

# We use Suds for Toolzilla SOAP API
suds-py3==1.3.3.0
# We use Zeep for contacting IPAM SOAP API
zeep==3.2.0

pymongo==3.7.2
pyyaml==3.13
uwsgi==2.0.17.1
requests[security]
requests_ntlm==1.1.0
tld==0.7.6
dnspython==1.16.0
python-ntlm3==1.0.2